% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Number_Formats.R
\name{nettle_bignumbers}
\alias{nettle_bignumbers}
\title{A function that formats big numbers
Numbers that have 12 digits are abreviated "T" for trillion
Numbers that have 9 digits are abreviated "B" for billion
Numbers that have 6 digits are abreviated "M" for Million
Commas are inserted}
\usage{
nettle_bignumbers(x)
}
\description{
A function that formats big numbers
Numbers that have 12 digits are abreviated "T" for trillion
Numbers that have 9 digits are abreviated "B" for billion
Numbers that have 6 digits are abreviated "M" for Million
Commas are inserted
}
\examples{
nettle_lib()
}

