% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Proper_Capwords.R
\name{nettle_capwords}
\alias{nettle_capwords}
\title{A function duplicating Excel's PROPER}
\usage{
nettle_capwords(s, strict = FALSE)
}
\description{
This function capitalizes the first letter of each word, but
TRIES to preserve acroynms it recognizes
}
\examples{
capwords(c("using AIC for model selection"))
capwords(c("using AIC", "for MODEL selection"), strict = TRUE)
}

