% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Nettle_Lib.R
\name{nettle_libs}
\alias{nettle_libs}
\title{A function that loads my favorite Hadley libraries
These are dplyr, tidyR, stringr, ggplot2, readxl, readr, lubridate, formatR}
\usage{
nettle_libs(x)
}
\description{
A function that loads my favorite Hadley libraries
These are dplyr, tidyR, stringr, ggplot2, readxl, readr, lubridate, formatR
}
\examples{
nettle_lib()
}

