% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Simple_Caps.R
\name{nettle_simplecaps}
\alias{nettle_simplecaps}
\title{A function duplicating Excel's PROPER}
\usage{
nettle_simplecaps(x)
}
\description{
This function capitalizes the first letter of each word and does NOT
preserve ACRONYMS
}
\examples{
nettle_simplecaps("the quick red fox jumps over the sleeping dog")
}

